Configurator now minds known minimal conflict sets
Adapted the pollution monitoring configurator so that it takes known
minimal conflict sets into account. The conflict sets are generated by
the genetic algorithm in this repository. This speeds up getting a
response in the case of a conflict by the factor 13 and creating a
diagnosis by the factor 6 on my development notebook. Pushed the version
number to 1.3.